+++ core/src/main/kotlin/com/jetbrains/aspire/actions/dashboard/resource/StopResourceAction.kt
@@ -1,9 +1,8 @@
 package com.jetbrains.aspire.actions.dashboard.resource
 
-import com.jetbrains.aspire.dashboard.StopResourceCommand
+import com.jetbrains.aspire.dashboard.findStopCommand
 import com.jetbrains.aspire.worker.AspireResourceData
 
-class StopResourceAction : ResourceCommandAction() {
-    override fun findCommand(resourceData: AspireResourceData) =
-        resourceData.commands.firstOrNull { it.name.equals(StopResourceCommand, true) }
+internal class StopResourceAction : ResourceCommandAction() {
+    override fun findCommand(resourceData: AspireResourceData) = resourceData.commands.findStopCommand()
 }